+++ package.json
@@ -3,8 +3,8 @@
   "version": "0.0.1",
   "main": "app.js",
   "scripts": {
-    "start": "NODE_ENV=production nodejs app.js",
-    "test": "PORT=3000 NODE_ENV=test nodejs app.js",
+    "start": "cross-env NODE_ENV=production node app.js",
+    "test": "cross-env PORT=3000 NODE_ENV=test node app.js",
     "build-front": "webpack --config webpack.config.js",
     "build-front-watch": "webpack --watch --config webpack.config.js"
   },
@@ -36,6 +36,7 @@
     "xmlbuilder": "^10.0.0"
   },
   "devDependencies": {
+    "cross-env": "^5.2.0",
     "pi-gpio": "0.0.8",
     "webpack": "^4.17.1",
     "webpack-cli": "^3.1.0"
