+++ public/js/src/map/index.js
@@ -1,6 +1,6 @@
 var livePoints = [];
-var liveMap = {};
-var liveBounds = {};
+var liveMap;
+var liveBounds;
 
 
 export default { livePoints, liveMap, liveBounds }
